requirements.txt
docker==4.4.0
PyYAML==5.3.1
kubernetes==12.0.1
-e git+https://github.com/rancher/client-python.git@v0.1.0
